requirements.txt
Flask
python-dotenv
Werkzeug
passkeys>=1.0.0